requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
exceptiongroup==1.1.1
Faker==18.3.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
wsproto==1.2.0
